# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Skybull Snapser CloudeCode",
    "description": "Skybull snapser cloud code",
    "contact": {},
    "version": "0.0.1"
  },
...
  "servers": [
    {
      "url": "//localhost:8080/"
    }
  ],
  "paths": {
    "/v1/byosnap-skybull-cloudcode/buy-energy": {
      "get": {
        "summary": "Get energy",
        "description": "Get energy",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Authorization Token",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "userId when they login",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Buy energy by gold",
        "description": "Buy energy by gold",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Authorization Token",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "userId when they login"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/byosnap-skybull-cloudcode/level-up-champion": {
      "post": {
        "summary": "Level up champion",
        "description": "Level up champion",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Authorization Token",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "sessionToken",
                  "userChampionUid",
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "userId when they login"
                  },
                  "sessionToken": {
                    "type": "string",
                    "description": "session token"
                  },
                  "userChampionUid": {
                    "type": "string",
                    "description": "champion userID"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}